chatbot-server/requirements.txt
Django
djangorestframework
daphne
channels
pycryptodomex
pycrypto
django-cors-headers